+++ packages/react-button/src/components/CompoundButton/renderCompoundButton.tsx
@@ -5,7 +5,7 @@ import type { CompoundButtonSlots, CompoundButtonState } from './CompoundButton.
 /**
  * Renders a CompoundButton component by passing the state defined props to the appropriate slots.
  */
-export const renderCompoundButton = (state: CompoundButtonState) => {
+export const renderCompoundButton_unstable = (state: CompoundButtonState) => {
   const { slots, slotProps } = getSlots<CompoundButtonSlots>(state);
   const { iconOnly, iconPosition } = state;