+++ packages/react-image/src/components/Image/useImage.ts
@@ -5,7 +5,7 @@ import type { ImageProps, ImageState } from './Image.types';
 /**
  * Given user props, returns state and render function for an Image.
  */
-export const useImage = (props: ImageProps, ref: React.Ref<HTMLImageElement>): ImageState => {
+export const useImage_unstable = (props: ImageProps, ref: React.Ref<HTMLImageElement>): ImageState => {
   const { bordered, fit, block, shape = 'square', shadow } = props;
 
   const state: ImageState = {