+++ packages/react-label/src/components/Label/useLabel.tsx
@@ -6,13 +6,13 @@ import { resolveShorthand } from '@fluentui/react-utilities';
 /**
  * Create the state required to render Label.
  *
- * The returned state can be modified with hooks such as useLabelStyles,
- * before being passed to renderLabel.
+ * The returned state can be modified with hooks such as useLabelStyles_unstable,
+ * before being passed to renderLabel_unstable.
  *
  * @param props - props from this instance of Label
  * @param ref - reference to root HTMLElement of Label
  */
-export const useLabel = (props: LabelProps, ref: React.Ref<HTMLElement>): LabelState => {
+export const useLabel_unstable = (props: LabelProps, ref: React.Ref<HTMLElement>): LabelState => {
   const { disabled = false, required = false, strong = false, size = 'medium' } = props;
   return {
     disabled,