+++ packages/react-menu/src/components/MenuList/useMenuListStyles.ts
@@ -1,4 +1,4 @@
-import { mergeClasses, makeStyles, shorthands } from '@fluentui/react-make-styles';
+import { mergeClasses, makeStyles, shorthands } from '@griffel/react';
 import type { MenuListState } from './MenuList.types';
 
 export const menuListClassName = 'fui-MenuList';
@@ -14,7 +14,7 @@ const useStyles = makeStyles({
 /**
  * Apply styling to the Menu slots based on the state
  */
-export const useMenuListStyles = (state: MenuListState): MenuListState => {
+export const useMenuListStyles_unstable = (state: MenuListState): MenuListState => {
   const styles = useStyles();
   state.root.className = mergeClasses(menuListClassName, styles.root, state.root.className);
   return state;