+++ packages/react-tooltip/src/components/Tooltip/useTooltipStyles.ts
@@ -1,4 +1,4 @@
-import { shorthands, makeStyles, mergeClasses } from '@fluentui/react-make-styles';
+import { shorthands, makeStyles, mergeClasses } from '@griffel/react';
 import { createArrowStyles } from '@fluentui/react-positioning';
 import { tokens } from '@fluentui/react-theme';
 import { arrowHeight } from './private/constants';
@@ -46,7 +46,7 @@ const useStyles = makeStyles({
 /**
  * Apply styling to the Tooltip slots based on the state
  */
-export const useTooltipStyles = (state: TooltipState): TooltipState => {
+export const useTooltipStyles_unstable = (state: TooltipState): TooltipState => {
   const styles = useStyles();
 
   state.root.className = mergeClasses(